backend/app/requirements.txt
fastapi
uvicorn
sqlmodel
requests
python-dotenv
anthropic>=0.40.0
httpx
